backend/requirements.txt
fastapi==0.111.0
pydantic==2.3.0
firebase-admin==6.4.0
passlib==1.7.4
python-jose==3.3.0
PyJWT==2.5.0
black==23.9.1
flake8==7.0.0
isort==5.13.2
mypy==1.5.1
pre-commit==3.7.1
sqlalchemy==2.0.27
sqlmodel==0.0.19
psycopg==3.1.18
psycopg-binary==3.1.19
psycopg-pool==3.2.1
psycopg2-binary==2.9.9
Mako==1.3.2
alembic==1.13.1
tenacity==8.2.3
python-multipart==0.0.9
uvicorn==0.27.1
email-validator==2.1.1
pydantic-settings==2.2.1
pytest>=7.4.2,<9.0.0
httpx>=0.25.0<0.26
bcrypt==4.0.1
